Create FloodLogic
I had to make a new account, not sure why it wouldn't let me sign in...
Anyway, this is the code that I wrote. Let me know if you have any questions. I ran it on a C compiler, it didn't work, I think it is because it tries to get data from outside of the array (ex: maze[i+1][j][4] = ?). Also remember that at the end of a column it jumps to the next row (ex: maze[2][0-1][4]=maze[1][15][4]). This code doesn't try to solve anything, it is just the flood logic. Good luck.