NEW SENSOR CODE!
Uses math based on the geometry of the IR light reflecting off of the
walls (at a certain angle) and being picked up by the IR sensor. The
sensor will pick up light intensity, based on the ratio of sensor area
to light area (a K/distance squared relationship). The K constant is
found, then used with preset distances to determine sensor values
(wallVal and motorVal). The code works rather well, but needs K ratios
for the different sensors, because not all sensors are identical.