Calibration Update
Attempted new calibration structure that uses delta values, with no
promising results. After taking measurements of the old code, putting
the data in excel, and using power fit lines, i noticed that the
lamda-distance relationship is not 1/x^2, but 1/x^p, where each sensor
has a different power (p). in the future, the calibration code should be
able to find the power of each sensor, because they seem to change
between runs. but for now these values are defined as static values base
on what the average value seemed to be.