Updated Sensor Curves (including diagonal)
The curves for the side and forward sensors are working well, but the
diagonal sensors are not. there seems to be a large amount of reflection
from the side and forward sensors affecting the diagonal sensor
readings. It is also likely that the diagonals are effecting the other
sensors, but due to the angle of reflection, it is not as noteworthy. A
new version of the sensor board may be necessary, with narrower led
projection width. Moving the diagonal sensors closer to the forward
sensor, but keeping the same 45 degree angle, would likely make it
easier for the diagonal sensors to detect pegs/walls sooner.
I also updated the distance values to detect walls, and changed the
maximum sensor value from 1026 to 800, because most sensors max out a
little higher than 800.