VC: Got I/O Working
Previously, when using interrupts based on the sampling rate, gathering
inputs and producing outputs at the same time wasn't working. Now, with
an interrupt timer library, I was able to get the input and output
working at the same time. By tweaking the chunk and multSlow variables,
it is now possible to reproduce a manipulated version of a speaker's
voice. VC_Main_2 is the current working version.