Saved VC_Main_3 as last-working-version, reworking program in VC_Main_4
VC_Main_3 works perfectly fine, but cannot change voice changer pitch in
real time. VC_Main_4 is attemting to add that functionality, but
requires significant code redesign. For example, it doesn't appear
possible to have the ADC sample voice and read the voltage of a POT, to
determine desired pitch. To remedy this, I am going into the ATMEGA328P
user's guide to get a better understanding of how to manipulate the ADC.